+++ src/pages/Mentos/EditMentos.jsx
@@ -1,7 +1,6 @@
 //  수정 페이지
 import { useEffect, useState } from "react";
 import { useParams } from "react-router-dom";
-import MainHeader from "@/components/MainHeader";
 import MentosForm from "./MentosForm.jsx";
 import { getMentos, updateMentos } from "@/services/mentos.api";
 
@@ -22,8 +21,7 @@ export default function EditMentosPage() {
   if (data === undefined) return <div className="p-6">불러오는 중…</div>;
   return (
     <div className="flex min-h-screen w-full justify-center overflow-x-hidden bg-[#f5f6f8] font-sans antialiased">
-      <section className="w-[375px] overflow-x-hidden rounded-xl bg-white px-4 py-5 shadow">
-        <MainHeader />
+      <section className="w-full overflow-x-hidden bg-white px-4 py-5 shadow">
         <h1 className="font-WooridaumB mt-6 mb-15 pl-2 text-[20px] font-bold">멘토링 수정하기</h1>
         <MentosForm mode="edit" initialValues={data ?? {}} onSubmit={handleUpdate} />
       </section>
